// archives/tdf-root-ui-legacy/api/types.ts
export type BandMemberDTO = {
  bmId: string;
  bmPartyId: number;
  bmPartyName: string;
  bmRole?: string | null;
};

export type BandDTO = {
  bandId: string;
  partyId: number;
  bName: string;
  bLabelArtist: boolean;
  bPrimaryGenre?: string | null;
  bHomeCity?: string | null;
  bPhotoUrl?: string | null;
  bContractFlags?: string | null;
  bMembers: BandMemberDTO[];
};